requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
neo4j>=5.20
groq>=0.9.0
feedparser==6.0.10
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0
deep-translator==1.11.4
schedule==1.2.0
pydantic==2.5.0
lxml==4.9.3
pillow==10.1.0